Allow toggling mouse capture and fix Action! tiny windows (#13)
Allow the user to toggle mouse capture using the F8 button.
Only init when the D3D create device call has a focus window. If the focus window is 0 then it hasn't come from the game. Action! was creating a device with a small resolution (not focused on a window) and the shim was being reinitialised and resized to this new small size.
Closes #12
Closes #11