Default to not on top and capturing mouse (#30)
Instead of having the `-notontop` flag have an `-ontop` flag and just disable it by default - this also addresses issue #29
Same with `-mousecaptured` -> `-nomousecapture`.
Dropped the `foreground` from `hide_cursor_foreground/show_cursor_foreground`
Removed the calls to `hide_cursor` in `WM_SETFOCUS` and replaced with `capture_mouse` as this keeps some other state consistent (and it already calls `hide_cursor` in it)
Removed the call to `show_cursor` in `WM_KILLFOCUS` as `release_mouse` already calls it.
Use `capture_mouse` instead of `toggle_capture` in the startup call.